docs: nextjs playground improve
Focus on server actions
