+++ playgrounds/nextjs/src/actions/auth.ts
@@ -1,13 +1,10 @@
+'use server'
+
 import { authed, pub } from '../orpc'
 import { CredentialSchema, TokenSchema } from '../schemas/auth'
 import { NewUserSchema, UserSchema } from '../schemas/user'
 
 export const signup = pub
-  .route({
-    method: 'POST',
-    path: '/signup',
-    summary: 'Sign up a new user',
-  })
   .input(NewUserSchema)
   .output(UserSchema)
   .func(async (input, context, meta) => {
@@ -19,11 +16,6 @@ export const signup = pub
   })
 
 export const signin = pub
-  .route({
-    method: 'POST',
-    path: '/signin',
-    summary: 'Sign in a user',
-  })
   .input(CredentialSchema)
   .output(TokenSchema)
   .func(async (input, context, meta) => {
@@ -33,11 +25,6 @@ export const signin = pub
   })
 
 export const refresh = authed
-  .route({
-    method: 'POST',
-    path: '/refresh',
-    summary: 'Refresh a token',
-  })
   .output(TokenSchema)
   .func(async (input, context, meta) => {
     return {
@@ -46,22 +33,12 @@ export const refresh = authed
   })
 
 export const revoke = authed
-  .route({
-    method: 'DELETE',
-    path: '/revoke',
-    summary: 'Revoke a token',
-  })
   .input(TokenSchema)
   .func(async (input, context, meta) => {
     // Do something
   })
 
 export const me = authed
-  .route({
-    method: 'GET',
-    path: '/me',
-    summary: 'Get the current user',
-  })
   .output(UserSchema)
   .func(async (input, context, meta) => {
     return context.user
