+++ playgrounds/nextjs/src/actions/planet.ts
@@ -1,3 +1,5 @@
+'use server'
+
 import { ORPCError } from '@orpc/server'
 import { oz } from '@orpc/zod'
 import { z } from 'zod'
@@ -10,28 +12,18 @@ import {
 } from '../schemas/planet'
 
 export const listPlanets = pub
-  .route({
-    method: 'GET',
-    path: '/',
-    summary: 'List all planets',
-  })
   .input(
     z.object({
       limit: z.number().int().min(1).max(100).optional(),
       cursor: z.number().int().min(0).default(0),
     }),
   )
-  .output(oz.openapi(z.array(PlanetSchema), { examples: [planets] }))
+  .output(z.array(PlanetSchema))
   .func(async (input, context, meta) => {
     return planets
   })
 
 export const createPlanet = authed
-  .route({
-    method: 'POST',
-    path: '/',
-    summary: 'Create a planet',
-  })
   .input(NewPlanetSchema)
   .output(PlanetSchema)
   .func(async (input, context, meta) => {
@@ -51,11 +43,6 @@ export const createPlanet = authed
   })
 
 export const findPlanet = pub
-  .route({
-    method: 'GET',
-    path: '/{id}',
-    summary: 'Find a planet',
-  })
   .input(
     z.object({
       id: z.number().int().min(1),
@@ -76,11 +63,6 @@ export const findPlanet = pub
   })
 
 export const updatePlanet = authed
-  .route({
-    method: 'PUT',
-    path: '/{id}',
-    summary: 'Update a planet',
-  })
   .input(UpdatePlanetSchema)
   .output(PlanetSchema)
   .func(async (input, context, meta) => {
@@ -101,11 +83,6 @@ export const updatePlanet = authed
   })
 
 export const updatePlanetImage = authed
-  .route({
-    method: 'PATCH',
-    path: '/{id}/image',
-    summary: 'Update a planet image',
-  })
   .input(
     z.object({
       id: z.number().int().min(1),
@@ -129,12 +106,6 @@ export const updatePlanetImage = authed
   })
 
 export const deletePlanet = authed
-  .route({
-    method: 'DELETE',
-    path: '/{id}',
-    summary: 'Delete a planet',
-    deprecated: true,
-  })
   .input(
     z.object({
       id: z.number().int().min(1),
