+++ playgrounds/nextjs/src/app/spec/route.ts
@@ -1,5 +1,5 @@
-import { router } from '@/router'
 import { generateOpenAPI } from '@orpc/openapi'
+import { router } from '../api/[...rest]/router'
 
 export function GET(request: Request) {
   const spec = generateOpenAPI({
@@ -9,6 +9,9 @@ export function GET(request: Request) {
       version: '1.0.0',
       description: `
 The example OpenAPI Playground for ORPC.
+
+> **Note**
+> This spec is generated without using \`.route()\` APIs, so you can see all methods is POST.
 
 ## Resources
